requirements.txt
asgiref == 3.8.1
Django == 5.1.1
djangorestframework == 3.15.2
pip == 23.2.1
psycopg2 == 2.9.9
sqlparse == 0.5.1
tzdata == 2024.1